Transferfunction now in its own window, interaction improved (ctrl-click for multiple selection), some zoom functionality added, boundingboxes tuned on points and lines. Quick update on preview-window to use gradient, not lagging anymore.

+++ include/inviwo/qt/editor/transferfunctioneditorcontrolpoint.h
@@ -58,7 +58,7 @@ namespace inviwo {
 
     private:
         int id_; ///< id to identify edge points
-        int size_; ///< size for drawing the points
+        float size_; ///< size for drawing the points
         TransferFunctionDataPoint* datapoint_; ///<The TransferFunctionDataPoint the controlpoint gets all its data from
     };
